Use the same utils file for the Uri extension [#572]
